requirements.txt
asgiref==3.3.1
Django==3.1.7
numpy==1.20.1
python-chess==0.31.3
pytz==2021.1
sqlparse==0.4.1
